requirements.txt
streamlit
load_dotenv
openai
extra_streamlit_components
